Remove .env file from repo.
The .env file was historically included in the repository as a template.
Although it is included in the .gitignore file, the .env file will still
be tracked by Git upon populating the file with actual credentials.
This introduces problems such as accidentally committing the file when
running commands like `git commit -am`. This change is necessary because
we are now expanding the project to other people. Testing will all be
done on the dev environment, and thus all developers will use the same
.env file which may be distributed centrally by project owners.