Added endpoints.js to store endpoint constants for API fetch calls.
endpoints.js is in src/constants. Since I made this branch from master,
there isn't a place where I can use the API's base url (an environment
variable) with an endpoint constant to make a fetch call(), so yeah
this is gonna be a short PR (I can do this in newSignupPage branch,
but the PR for that is already so bloated I don't want to add more).

And also, to be frank, the biggest change for this PR is in .env, so
that's also why this is such a short PR.