Added import/order rules to auto rearrange our imports.
I ran 'npm run lint' but didn't really see things change. Idk if
that's just because the files I checked were already in the correct
import style, or it just didn't work... Also, had to put it in
"settings" instead of having it be a top-level "import/order" rule,
even though "eslint-plugin-import" docs on "import/order" seems
to indicate that it could be top-level.